Added the `Secure` annotation
This update allows you to do something like what is described at the
bottom of http://silex.sensiolabs.org/doc/providers/security.html
except with annotations.

I added a symfony/security suggest in composer.json to indicate that the
Symfony Security component is required to use this annotation
successfully.

I removed the composer.lock file because is is superfulous in a library
project.